quantum-annealing-project/pyproject.toml
[project]
name = "quantum-annealing-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dwave-ocean-sdk>=8.4.0",
    "matplotlib>=3.10.3",
    "openjij>=0.10.14",
    "pyqubo>=1.5.0",
    "optuna>=3.0.0",
]
